Add some details about TAP metadata
Describe the two types of TAP schema metadata besides the schema
itself that are currently published. Expand on the plan for using
service discovery to locate the current versions of those files,
and using a `versions` table in the TAP schema database to indicate
what version of the TAP schema is loaded.